requirements.txt
openai
python-dotenv
streamlit
fpdf
Py2PDF
pptx